+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed Wrench in Dismantling Mode voiding Storage Block contents (raoulvdberge)
 - Fixed OPs not having global permissions on secured storage systems (raoulvdberge)
 - Fixed crash when Destructor tries to break secured network block (raoulvdberge)
+- Fixed Fluid Interface not dropping inventory contents (raoulvdberge)
 
 ### 1.4
 - Added Security Manager (raoulvdberge)